pyproject.toml
[project]
name = "web-search-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai>=0.6.3",
    "fastmcp>=2.9.2",
    "googlesearch-python>=1.3.0",
    "pydantic>=2.11.7",
]
